NHGH-2558 EDI SQL Updates

+++ nhgh-edi-backend/db-edi/usbdg/Tables/event.sql
@@ -6,12 +6,14 @@
     [zutc_now]      DATETIME2 (7)  NOT NULL,
     [zcell_info]    VARCHAR (100)  NULL,
     [zbatt_volt]    SMALLINT       NULL,
-    [zbatt_chrg]    TINYINT        NULL,
+    [zbatt_chrg]    SMALLINT       NULL,
     [DATEADDED]     DATETIME2 (7)  NULL,
     [TAMB]          DECIMAL (3, 1) NULL,
     PRIMARY KEY CLUSTERED ([zutc_now] ASC, [ESER] ASC),
     CONSTRAINT [FK_usbdg_event.ESER] FOREIGN KEY ([ESER]) REFERENCES [usbdg].[device] ([ESER])
 );
+
+
